requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
distro==1.9.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.0
markdown-it-py==4.0.0
mdurl==0.1.2
openai==2.3.0
pydantic==2.12.0
pydantic_core==2.41.1
Pygments==2.19.2
python-dotenv==1.1.1
rich==14.2.0
smmap==5.0.2
sniffio==1.3.1
tqdm==4.67.1
tree-sitter==0.25.2
typing-inspection==0.4.2
typing_extensions==4.15.0
